retrieval/services/database.py
"""
Database service

Handles interactions with MySQL database for paper metadata.
"""

from typing import List, Optional

import mysql.connector
from urllib.parse import urlparse

from retrieval.config import settings

# TODO: Implement database service
# - Connect to MySQL
# - Retrieve paper metadata
# - Query papers with filters
# - Get paper details


class DatabaseService:
    """Service for MySQL database operations"""

    def __init__(self):
        """Initialize database connection"""
        self.connection = None

    def get_connection(self):
        """Get database connection"""
        if self.connection is None or not self.connection.is_connected():
            self.connection = self._create_connection()
        return self.connection

    def _create_connection(self):
        """Create database connection"""
        if settings.DATABASE_URL:
            parsed = urlparse(settings.DATABASE_URL)
            db_config = {
                "host": parsed.hostname or settings.DB_HOST,
                "port": int(parsed.port) if parsed.port else int(settings.DB_PORT),
                "database": parsed.path.lstrip("/") if parsed.path else settings.DB_NAME,
                "user": parsed.username or settings.DB_USER,
                "password": parsed.password or settings.DB_PASSWORD,
            }
        else:
            db_config = {
                "host": settings.DB_HOST,
                "port": int(settings.DB_PORT),
                "database": settings.DB_NAME,
                "user": settings.DB_USER,
                "password": settings.DB_PASSWORD,
            }

        return mysql.connector.connect(**db_config)

    def get_paper_by_id(self, paper_id: str) -> Optional[dict]:
        """
        Get paper metadata by ID.

        Args:
            paper_id: Paper ID (arxiv_id or database id)

        Returns:
            Paper metadata dictionary or None
        """
        # TODO: Implement paper retrieval
        raise NotImplementedError("Paper retrieval not yet implemented")

    def list_papers(
        self,
        limit: int = 20,
        offset: int = 0,
        category: Optional[str] = None,
    ) -> List[dict]:
        """
        List papers with pagination and optional filtering.

        Args:
            limit: Maximum number of results
            offset: Pagination offset
            category: Optional category filter

        Returns:
            List of paper metadata dictionaries
        """
        # TODO: Implement paper listing
        raise NotImplementedError("Paper listing not yet implemented")

    def close(self):
        """Close database connection"""
        if self.connection and self.connection.is_connected():
            self.connection.close()

